requirements.txt
medpy
matplotlib
numpy
plaidml
plaidml-keras
opencv-python
scipy